collision: merge Collide_GetCollisionInfo to trx
