+++ src/tr1/game/collide.h
@@ -3,10 +3,6 @@
 #include <libtrx/game/collision.h>
 #include <libtrx/game/items/types.h>
 
-bool Collide_CollideStaticObjects(
-    COLL_INFO *coll, int32_t x, int32_t y, int32_t z, int16_t room_num,
-    int32_t height);
-
 int32_t Collide_GetSpheres(ITEM *item, SPHERE *slist, int32_t world_space);
 
 int32_t Collide_TestCollision(ITEM *item, ITEM *lara_item);
