+++ src/tr2/game/collide.h
@@ -3,10 +3,6 @@
 #include <libtrx/game/collision.h>
 #include <libtrx/game/items/types.h>
 
-bool Collide_CollideStaticObjects(
-    COLL_INFO *coll, int32_t x, int32_t y, int32_t z, int16_t room_num,
-    int32_t height);
-
 int32_t Collide_TestCollision(ITEM *item, const ITEM *lara_item);
 
 int32_t Collide_GetSpheres(const ITEM *item, SPHERE *spheres, bool world_space);
